requirements.txt
requests>=2.31.0
kafka-python>=2.0.2
python-dotenv>=1.0.0
streamlit>=1.28.1
pandas>=2.0.3
apache-airflow==2.7.3
apache-airflow-providers-apache-kafka==5.1.2